Implemented check to prevent user from updating another user contact

+++ controllers/contact_controller.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler")
 const Contact = require("../models/contactModel")
 
 const getContacts= asyncHandler(async (req,res)=>{
-    const contacts = await Contact.find()
+    const contacts = await Contact.find({user_id:req.user._id})
     res.status(200).json(contacts)
 })
 
@@ -13,10 +13,12 @@ const createContact= asyncHandler(async (req,res)=>{
         res.status(400)
         throw new Error("All Fields are manadatory")
     }
+
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id:req.user._id
     })
 
     res.status(201).json(contact)
@@ -28,6 +30,12 @@ const updateContact= asyncHandler(async (req,res)=>{
         res.status(404)
         throw new Error("Contact not found")
     }
+
+    if(contact.user_id.toString() !== req.user._id){
+        res.status(403)
+        throw new Error("User dont have permission to update another user contact")
+    }
+
     const updated_contact = await Contact.findByIdAndUpdate(
         contact.id,
         req.body,
@@ -42,6 +50,12 @@ const deleteContact= asyncHandler(async (req,res)=>{
         res.status(404)
         throw new Error("Contact not found")
     }
+
+    if(contact.user_id.toString() !== req.user._id){
+        res.status(403)
+        throw new Error("User dont have permission to delete another user contact")
+    }
+
     await Contact.deleteOne({ _id: req.params.id })
     res.status(200).json(contact)
 })
